==> notebooks/2_preprocess_tweets.ipynb <==
import pyspark, pickle
from pyspark import SparkContext
from pyspark.sql.functions import countDistinct
from pyspark.storagelevel import StorageLevel
import pandas as pd
import numpy as np

spark = pyspark.sql.SparkSession.builder.getOrCreate()
sc = spark.sparkContext

## Check schema of dataframe made from tweets

df = spark.read.json('tweet_storage1/')
df.printSchema()

## Function to read in tweet files, extract necessary data, and save

def extract_tweet_data(tweets_folder, save_as):
    """
    tweets_folder: a folder with .json files containing tweets
    save_as: filename to save processed tweets
    This function loads the tweets, extracts the necessary data, and saves a new parquet file with that data.
    """

    df = spark.read.json(tweets_folder)
    df.registerTempTable("df")

    # Extract tweets, nested retweets, and nested quoted tweets into separate flattened dataframes
    flat_tweets = spark.sql("""
        select id_str as tweet_id, coordinates.coordinates[0] as lon, coordinates.coordinates[1] as lat, created_at,
            in_reply_to_status_id_str as in_reply_to_tweet_id, in_reply_to_screen_name, source, user.description,
            user.followers_count, user.friends_count, user.id_str as user_id, user.screen_name as screen_name,
            user.location, user.name, user.statuses_count, user.time_zone, user.utc_offset, retweet_count, 
            retweeted_status.id_str as retweet_id, retweeted_status.user.screen_name as retweeted_screen_name,
            quoted_status_id_str as quote_id, quoted_status.user.screen_name as quoted_screen_name, entities.hashtags,
            entities.user_mentions,
                case when retweeted_status is not null then retweeted_status.text
                else text end as text

        from df
    """)

    flat_RTs = spark.sql("""
        select id_str as tweet_id, coordinates.coordinates[0] as lon, coordinates.coordinates[1] as lat, created_at,
            in_reply_to_status_id_str as in_reply_to_tweet_id, in_reply_to_screen_name, source, user.description,
            user.followers_count, user.friends_count, user.id_str as user_id, user.screen_name as screen_name,
            user.location, user.name, user.statuses_count, user.time_zone, user.utc_offset, retweet_count, 
            null as retweet_id, null as retweeted_screen_name, quoted_status_id_str as quote_id,
            quoted_status.user.screen_name as quoted_screen_name, entities.hashtags, entities.user_mentions, text
        from 
            (select retweeted_status.*
             from df
             where retweeted_status is not null
            ) as sub
    """)

    flat_QTs = spark.sql("""
        select id_str as tweet_id, coordinates.coordinates[0] as lon, coordinates.coordinates[1] as lat, created_at,
            in_reply_to_status_id_str as in_reply_to_tweet_id, in_reply_to_screen_name, source, user.description,
            user.followers_count, user.friends_count, user.id_str as user_id, user.screen_name as screen_name,
            user.location, user.name, user.statuses_count, user.time_zone, user.utc_offset, retweet_count, 
            null as retweet_id, null as retweeted_screen_name, quoted_status_id_str as quote_id,
            null as quoted_screen_name, entities.hashtags, entities.user_mentions, text
        from 
            (select quoted_status.*
             from df
             where quoted_status is not null
            ) as sub
    """)

    flat_tweets.registerTempTable('flat_tweets')
    flat_RTs.registerTempTable('flat_RTs')
    flat_QTs.registerTempTable('flat_QTs')
    
    # Combine the above and keep distinct tweet_ids
    undistinct_tweets = spark.sql("""
        select *
        from flat_tweets
        union all
        select *
        from flat_RTs
        union all
        select *
        from flat_QTs

    """)
    undistinct_tweets.registerTempTable('undistinct_tweets')
    
    distinct_tweets = spark.sql("""
        select *
        from
            (select *,
                row_number() over (partition by tweet_id order by created_at) as n_repeats
            from undistinct_tweets) as sub
        where n_repeats = 1
    """)
    distinct_tweets.persist(StorageLevel.MEMORY_AND_DISK)
    distinct_tweets.registerTempTable('distinct_tweets')
    
    distinct_tweets.write.parquet(save_as)
    print('Extracted and saved!')
    
    

## Extract and save data

extract_tweet_data('tweet_storage1/', 'tweets1.parquet')

extract_tweet_data('tweet_storage2/', 'tweets2.parquet')

## Load simplified data and union distinct tweet_ids

tweets1 = spark.read.parquet('tweets1.parquet')
tweets2 = spark.read.parquet('tweets2.parquet')

tweets1.registerTempTable('tweets1')
tweets2.registerTempTable('tweets2')

all_tweets = spark.sql("""
    select *
    from
        (select *,
            row_number() over (partition by tweet_id order by created_at) as n_repeats2
        from
            (select *
            from tweets1

            union all

            select *
            from tweets2) sub
        ) sub2
    where n_repeats2 = 1
""")
all_tweets.persist(StorageLevel.MEMORY_AND_DISK)
all_tweets.registerTempTable('all_tweets')

all_tweets.count()

## Do some checks on the data

# View some of the data
all_tweets.filter('lon is not null').select('tweet_id', 'screen_name', 'lon', 'lat', 'text').show()

# Number of distinct users: 306,753
all_tweets.agg(countDistinct(all_tweets.user_id)).show()

# Most retweeted users
spark.sql("""
    select screen_name, sum(retweet_count) as total_rts
    from all_tweets
    where retweet_id is null
    group by screen_name
    order by total_rts desc
""").show()

# Users with most followers
spark.sql("""
    select screen_name, max(followers_count) as n_followers
    from all_tweets
    group by screen_name
    order by n_followers desc
""").show()

## Read a bunch of tweets from each of my search categories to make sure they are generally relevant

# 'climatechange' is all about climate change

spark.sql("""
    select screen_name, created_at, text
    from all_tweets
    where text like '%climatechange%'
""").take(50)

# '#climate' is all about climate change

spark.sql("""
    select screen_name, created_at, text
    from all_tweets
    where text like '%#climate %'
""").take(50)

# 'climate' and combinations of certain other words are all about climate change

spark.sql("""
    select screen_name, created_at, text
    from all_tweets
    where text like '%climate%' and text rlike 'science|scientists?|alarmists?|change|realists?|denial|deniers?' 
""").take(50)

# 'globalwarming' and 'global' or 'warming' are all about climate change

spark.sql("""
    select screen_name, created_at, text
    from all_tweets
    where (text like '%globalwarming%') or (text like '%global%' and text like '%warming%') 
""").take(50)

# 'agw' grabs lots of irrelevant stuff (mostly in Japanese...)
# I changed my API query to search specifically for '#agw', which appears to be much more effective

spark.sql("""
    select screen_name, created_at, text
    from all_tweets
    where text like '%#agw%'
""").take(50)

## Delete tweets that reference Pitbull's new album, "Climate Change"

delete_tweets = spark.sql("""
    select distinct tweet_id
    from 
        (select tweet_id
        from all_tweets
        where screen_name in ('pitbull', 'RCARecords', 'Genius') or
            retweeted_screen_name in ('pitbull', 'RCARecords', 'Genius') or
            quoted_screen_name in ('pitbull', 'RCARecords', 'Genius') or
            in_reply_to_screen_name in ('pitbull', 'RCARecords', 'Genius')

        union

        select tweet_id
        from
            (select tweet_id, explode(user_mentions) as user_mentions
            from all_tweets) sub
        where user_mentions.screen_name in ('pitbull', 'RCARecords', 'Genius')
        ) delete_tweets_sub
""")
delete_tweets.registerTempTable('delete_tweets')

all_tweets = spark.sql("""
    select a.*
    from all_tweets as a
    left join delete_tweets as d
    on a.tweet_id = d.tweet_id
    where d.tweet_id is null
""")
all_tweets.registerTempTable('all_tweets')

all_tweets.count()

## Remove bots

# View retweet/tweet ratios for users. 

spark.sql("""
select screen_name, n_rt, n_orig, n_rt/(n_rt + n_orig)*100 as rt_tw_ratio
from
    (select screen_name, 
        sum(case when type = 'rt' then n else 0 end) as n_rt,
        sum(case when type = 'orig' then n else 0 end) as n_orig
    from
        (select screen_name, type, count(*) as n
        from
            (select screen_name,
                case when retweet_id is null then 'orig'
                else 'rt' end as type
            from all_tweets) as sub
        group by screen_name, type) as sub2
    group by screen_name) as sub3
order by rt_tw_ratio desc, n_rt desc
""").show()

#Remove users that have retweet/tweet ratios greater than 97%, and that have more than 10 retweets.
# They are either bots or curated lists who don't engage in conversation.

bot_list = spark.sql("""
    select screen_name
    from
        (select screen_name, 
            sum(case when type = 'rt' then n else 0 end) as n_rt,
            sum(case when type = 'orig' then n else 0 end) as n_orig
        from
            (select screen_name, type, count(*) as n
            from
                (select screen_name,
                    case when retweet_id is null then 'orig'
                    else 'rt' end as type
                from all_tweets) as sub
            group by screen_name, type) as sub2
        group by screen_name
        having n_rt/(n_rt + n_orig)*100 > 97 and n_rt > 10) bots
""")
bot_list.registerTempTable('bot_list')
bot_list.persist(StorageLevel.MEMORY_AND_DISK)

bot_list.count()

all_tweets = spark.sql("""
    select t.screen_name as screen_name, tweet_id, lon, lat, created_at, in_reply_to_tweet_id, in_reply_to_screen_name, source, description,
        followers_count, friends_count, user_id, location, name, statuses_count, time_zone, utc_offset, retweet_count,
        retweet_id, retweeted_screen_name, quote_id, quoted_screen_name, hashtags, user_mentions, text
    from all_tweets t
    left join bot_list b
    on b.screen_name = t.screen_name
    where b.screen_name is null
""")

all_tweets.count()

## Save tweets

all_tweets.write.parquet('tweets_all.parquet')

## Extract the messy nested objects (i.e. hashtags and user_mentions) from the tweets  
## Save as separate tables

hashtags = spark.sql("""
    select distinct tweet_id, lower(hashtags.text) as hashtag
    from
        (select tweet_id, explode(hashtags) as hashtags
        from all_tweets) as sub
""")

hashtags.show()

hashtags.write.parquet('hashtags_all.parquet')

mentions = spark.sql("""
    select distinct tweet_id, mentions.screen_name as mention_screen_name
    from
        (select tweet_id, explode(user_mentions) as mentions
        from all_tweets) as sub
""")

mentions.show()

mentions.write.parquet('mentions_all.parquet')

## Check on tweets that I might be missing  
### (I didn't end up doing anything with this.)

# Of my tweets, 18067 are replies to distinct statuses that I didn't download
tweet_ids_to_get = spark.sql("""
    select distinct tweet_id
    from
        (select in_reply_to_tweet_id as tweet_id
        from all_tweets
        except
        select tweet_id
        from all_tweets) as sub
""").collect()

tweet_ids_to_get = [str(x.tweet_id) for x in tweet_ids_to_get]

# use tweepy API.statuses_lookup with include_entities set to True
# http://docs.tweepy.org/en/v3.5.0/api.html

# with open('tweet_ids_to_get.pkl', 'wb') as picklefile:
#     pickle.dump(tweet_ids_to_get, picklefile)

## Notes on what some of the tweet fields mean

# Showing that if you retweet, your retweet_count shows the number of RTs of the original tweet
df.select('user.screen_name', 'text', 'retweet_count', 'retweeted_status', 'retweeted_status.user.screen_name', 
          'retweeted_status.text').limit(20).show()

# Showing, with the above table for reference, that quoted tweets don't count as RTs
df.filter('quoted_status is not null').select('user.id_str', 'user.screen_name', 'text',
                'quoted_status', 'quoted_status.user.screen_name', 'quoted_status.text').limit(20).show()

# Showing that even if your tweet is a quote, you can still get RTs on it
df.filter('quoted_status is not null and retweet_count > 0').select('user.screen_name', 'text',
            'quoted_status.user.screen_name', 'quoted_status.text', 'retweet_count').limit(20).show()
