==> notebooks/3_model_climate_acceptance.ipynb <==
import pyspark, pickle
from pyspark import SparkContext
from pyspark.sql.functions import countDistinct
from pyspark.storagelevel import StorageLevel
import pandas as pd
import numpy as np
from pyspark.ml.feature import CountVectorizer, StringIndexer, StopWordsRemover, NGram, RegexTokenizer

from nltk.corpus import stopwords
import nltk, re

from pyspark.ml import PipelineModel
from pyspark.ml.evaluation import MulticlassClassificationEvaluator, BinaryClassificationEvaluator
from pyspark.ml.tuning import ParamGridBuilder, CrossValidator, CrossValidatorModel
from pyspark.mllib.evaluation import MulticlassMetrics
from pyspark.ml.classification import NaiveBayes
from pyspark.ml import Pipeline

from pyspark.sql.types import StringType

from sklearn import svm, grid_search, datasets
from spark_sklearn import GridSearchCV

pd.options.display.max_colwidth = -1

spark = pyspark.sql.SparkSession.builder.getOrCreate()
sc = spark.sparkContext

## Read data and clean

tweet_data = pd.read_csv('labeled_tweets.csv')

# Clean tweet_data

tweet_data.columns = ['garbage', 'tweet', 'existence', 'existence_conf']
tweet_data.drop('garbage', axis=1, inplace=True)
tweet_data.existence.replace(['Yes', 'No', 'Y', 'N'], ['yes', 'no', 'yes', 'no'], inplace=True)
tweet_data.tweet = tweet_data.tweet.str.replace('https?://[^ ,]+', '[link]') #replace links with '[link]'
tweet_data.existence.fillna('neutral', inplace=True)
tweet_data = tweet_data.sort_values('existence_conf', ascending=False).drop_duplicates(subset=['tweet'])
tweet_data = tweet_data[tweet_data.tweet != '[link]']   # One tweet consists only of a link

# Remove 'neutral' tweets. Will build model using only 'accept' and 'deny' classes
tweet_data = tweet_data[(tweet_data.existence != 'neutral')]

print(tweet_data.shape)
tweet_data.head()

# Check class balance
tweet_data.existence.value_counts(dropna=False, normalize=True)

## Convert to Spark dataframe and do train/test split

# Create spark dataframe of tweets called 'tweets spark'

tweet_sp = spark.createDataFrame(tweet_data)
tweet_sp.persist()
tweet_sp.show()

(train, test) = tweet_sp.randomSplit([0.7, 0.3], seed = 100)

train.persist()
test.persist()

print(train.count())
print(test.count())

# Naive Bayes model
## Set parameters

# Create regex tokenizer that is useful for Twitter data (preserves emoticons, hashtags, etc.)
# I used code from here, with some modifications: https://github.com/adonoho/TweetTokenizers/blob/master/PottsTweetTokenizer.py

# I got rid of ellipsis matcher
# I modified words with apostrophes to not keep after the apostrophe
# I got rid of the 'everything else that isn't whitespace' matcher
# I added regex to capture words in quotes as separate items: r'(?:["\'][a-zA-Z0-9/-]+["\'])'
pattern = r"""(?:\[link\])|(?:(?:\+?[01][\-\s.]*)?(?:[\(]?\d{3}[\-\s.\)]*)?\d{3}[\-\s.]*\d{4})|(?:(?<= )[<>]?[:;=8][\-o\*\']?[\)\]\(\[dDpP/\:\}\{@\|\\]|[\)\]\(\[dDpP/\:\}\{@\|\\][\-o\*\']?[:;=8][<>]?)|(<[^>]+>)|(?:@[\w_]+)|(?:\#+[\w_]+[\w\'_\-]*[\w_]+)|(?:["\'][a-z0-9/-]+["\'])|(?:[a-z][a-z\-_]+[a-z])|(?:[+\-]?\d+[,/.:-]\d+[+\-]?)|(?:[\w_]+)"""

word_re = re.compile(pattern, re.VERBOSE | re.I | re.UNICODE)

# Test Regex
df = spark.createDataFrame(["RT @contessabrewer: @newsbusters That's not what I said. >:(  :D I said that snowstorms don't refute global warming. (RE: [link]",
                           "@OTOOLEFAN REAL science. Not Algore's climate change 'science' hacks with their man-made global warming hoax. Gore=Palin"],
                           StringType())
df.show()

RegexTokenizer(inputCol="value", outputCol="tokens", gaps=False, pattern=word_re.pattern).transform(df).take(2)

# Use this to check how labels are indexed by StringIndexer

#StringIndexer(inputCol="existence", outputCol="label").fit(train).transform(train).show()

# Number my labels
# Label numbering goes in order of most frequent label, descending
label_indxr = StringIndexer(inputCol="existence", outputCol="label")

# Tokenize tweets
tokenizer = RegexTokenizer(inputCol="tweet", outputCol="tokens", gaps=False, pattern=word_re.pattern)

# Remove stopwords
stp_rmv = StopWordsRemover(inputCol=tokenizer.getOutputCol(), outputCol='new_tokens',
                           stopWords=stopwords.words('english'))

# Ngrams
ngram = NGram(inputCol=stp_rmv.getOutputCol(), outputCol="ngrams")

# Count occurences of words
cnvk = CountVectorizer(inputCol=ngram.getOutputCol(), outputCol='counts')

# Train a NaiveBayes model
nb = NaiveBayes(featuresCol=cnvk.getOutputCol(), modelType="multinomial")

# Pipeline
pipeline = Pipeline(stages=[label_indxr, tokenizer, stp_rmv, ngram, cnvk, nb])

# Create evaluator
evaluator = MulticlassClassificationEvaluator(labelCol="label", predictionCol="prediction",
                                             metricName="accuracy")
# evaluator = BinaryClassificationEvaluator(labelCol='label',
#                                         rawPredictionCol='prediction',
#                                         metricName='areaUnderROC')

# Program search params
param_grid = (ParamGridBuilder() 
    .addGrid(nb.smoothing, [0.0, 0.2, 0.4, 0.6, 0.8, 1.0]) 
    .addGrid(ngram.n, [1, 2, 3]) \
    .build())

# Put pipeline together with param search
cv_pipe = CrossValidator(estimator=pipeline, estimatorParamMaps=param_grid, evaluator=evaluator)

## Train model and evaluate

# Train model
nb_model = cv_pipe.fit(train)

# Average accuracy across grid search:
nb_model.avgMetrics

# Show each model param alongside accuracy
list(zip(param_grid, nb_model2.avgMetrics))

# See accuracy on test set
test_prediction = nb_model.transform(test)

accuracy = evaluator.evaluate(test_prediction)
print("Model Accuracy: ", accuracy)

metrics = MulticlassMetrics(test_prediction.select("prediction", "label").rdd)

print('Yes recall:', metrics.recall(0))
print('No recall:', metrics.recall(1), '\n')

print('Yes precision:', metrics.precision(0))
print('No precision:', metrics.precision(1))

# Show normalized confusion matrix
conf = metrics.confusionMatrix().toArray()
conf_norm = conf/conf.sum(axis=1)[:, np.newaxis]
print(conf_norm)

## Save naive bayes model pipeline

nb_model.bestModel.save('./nb_model_pipeline')

#model = PipelineModel.load('./nb_model_pipeline/')

# SVM Model
## Set parameters

# Number my labels
# Label numbering goes in order of most frequent label, descending
label_indxr_svm = StringIndexer(inputCol="existence", outputCol="label")

# Tokenize tweets
tokenizer_svm = RegexTokenizer(inputCol="tweet", outputCol="tokens", gaps=False, pattern=word_re.pattern)

# Remove stopwords
stp_rmv_svm = StopWordsRemover(inputCol=tokenizer_svm.getOutputCol(), outputCol='new_tokens',
                           stopWords=stopwords.words('english'))

# Count occurences of words
cnvk_svm = CountVectorizer(inputCol=stp_rmv_svm.getOutputCol(), outputCol='counts')

# Pipeline
pipeline_svm = Pipeline(stages=[label_indxr_svm, tokenizer_svm, stp_rmv_svm, cnvk_svm])

# Create evaluator
evaluator_svm = MulticlassClassificationEvaluator(labelCol="label", predictionCol="prediction",
                                             metricName="f1")

## Train model  
## Pyspark SVM doesn't have rbf kernel, so I have to use spark_sklearn library

# Count Vectorize my data
countvec_data_svm = pipeline_svm.fit(train).transform(train)
countvec_data_svm.show()

# Split up the sparse matrix that Spark gives me and create a standard sklearn X matrix
countvec_data_svm_x = pd.DataFrame(countvec_data_svm.select('counts').rdd.map(
                        lambda x: x['counts'].toArray()).collect())

# Create standard sklearn Y vector
countvec_data_svm_y = countvec_data_svm.select('existence').toPandas().existence

# Train model
parameters = [{'kernel': ['rbf'], 'gamma': [1,1e-1,1e-2,1e-3, 1e-4], 'C': [1, 10, 100, 1000]},
                    {'kernel': ['linear'], 'C': [1, 10, 100, 1000]}]

svr = svm.SVC()
clf = GridSearchCV(sc, svr, parameters)
clf.fit(countvec_data_svm_x, countvec_data_svm_y)

## Evaluate model based on CV accuracy, and test accuracy

# CV accuracy:
clf.best_score_

clf.best_params_

# Run test data through pipeline
test_svm = pipeline_svm.fit(train).transform(test)
test_svm_x = pd.DataFrame(test_svm.select('counts').rdd.map(
                        lambda x: x['counts'].toArray()).collect())
test_svm_y = test_svm.select('existence').toPandas().existence.to_frame()

# Make predictions
svm_predicts = clf.predict(test_svm_x)

# Compare predictions and ground truth
test_svm_y['predict'] = svm_predicts
test_svm_y.head()

# Recall for each class
test_svm_y.groupby('existence').predict.value_counts(normalize=True)

# Precision for each class
test_svm_y.groupby('predict').existence.value_counts(normalize=True)

SVM performs worse than Naive Bayes here, so I won't keep it
